login page done

+++ server/routes/api/Account.js
@@ -3,11 +3,11 @@ const router = express.Router();
 
 const Account = require("../../schemas/Account");
 
-router.get("/test", (req, res) => {
-  console.log("Hit at http://localhost:3001/api/account/test");
-  res.json({
-    msg: "This is the Account route.",
-  });
+router.get("/allAccounts", (req, res) => {
+  console.log("Hit at http://localhost:3001/api/account/allAccounts");
+  Account.find()
+    .then((acc) => res.json(acc))
+    .catch((err) => res.status(404).json({ account_not_found: "No accounts" }));
 });
 
 router.get("/login/:email", (req, res) => {
@@ -15,7 +15,7 @@ router.get("/login/:email", (req, res) => {
   console.log(`Hit at http://localhost:3001/api/account/login/${email}`);
   Account.findOne({ email: email })
     .then((acc) => {
-      res.json(acc);
+      return res.json(acc);
     })
     .catch((err) =>
       res
@@ -24,30 +24,36 @@ router.get("/login/:email", (req, res) => {
     );
 });
 
-router.get("/allAccounts", (req, res) => {
-  console.log("Hit at http://localhost:3001/api/account/allAccounts");
-  Account.find()
-    .then((acc) => res.json(acc))
-    .catch((err) => res.status(404).json({ account_not_found: "No accounts" }));
-});
-
 router.post("/signup", (req, res) => {
   console.log("Hit at http://localhost:3001/api/account/signup");
 
   const newAccount = new Account({
-    id: req.body.id,
     email: req.body.email,
     password: req.body.password,
   });
 
-  newAccount.save().then((newAcc) => res.json(newAcc));
+  Account.exists({ email: req.body.email }, (err, doc) => {
+    if (doc) {
+      res.json({ success: false });
+    } else {
+      newAccount.save().then((newAcc) => res.json(newAcc));
+    }
+  });
 });
 
 router.delete("/delete", (req, res) => {
   console.log("Hit at http://localhost:3001/api/account/delete");
   let email = req.body.email;
   let password = req.body.password;
-  Account.findOneAndRemove({ email: email, password: password }).exec();
+
+  Account.exists({ email: email, password: password }, (err, doc) => {
+    if (doc) {
+      Account.findOneAndRemove({ email: email, password: password }).exec();
+      res.json({ success: true });
+    } else {
+      res.json({ success: false });
+    }
+  });
 });
 
 module.exports = router;
